Gutenberg: modify BrowserURL to match Calypso routes
